Resolves: update framework
(https://github.com/musescore/framework_tmp/pull/29)

<!-- Use "x" to fill the checkboxes below like [x] -->

- [x] I signed [CLA](https://www.audacityteam.org/cla/)
- [x] The title of the pull request describes an issue it addresses
- [x] If changes are extensive, then there is a sequence of easily
reviewable commits
- [x] Each commit's message describes its purpose and effects
- [x] There are no behavior changes unnecessary for the stated purpose
of the PR

Recommended:
- [ ] Each commit compiles and runs on my machine without known
undesirable changes of behavior

QA:

- [ ] Autobot test cases have been run

+++ src/app/guiapp.cpp
@@ -429,9 +429,9 @@ void GuiApp::destroyContext(const modularity::ContextPtr& ctx)
     modularity::removeIoC(ctx);
 }
 
-int GuiApp::contextCount() const
+size_t GuiApp::contextCount() const
 {
-    return static_cast<int>(m_contexts.size());
+    return m_contexts.size();
 }
 
 std::vector<modularity::ContextPtr> GuiApp::contexts() const
